application put into a modular component, still reliant on connections prop from index.jsx file
